Refactor `ApiManager`
Its config is now backend-independent.
